+++ day-7/main.py
@@ -1,5 +1,5 @@
-from src.commons.bet_winning_manager import SimpleBetManager as FirstBetManager
-from src.commons.bet_winning_manager import BetWithJokerManager as SecondBetManager
+from src.part_1.simple_bet_manager import SimpleBetManager as FirstBetManager
+from src.part_2.bet_with_joker_manager import BetWithJokerManager as SecondBetManager
 from src.bet_winnings_calculator import WinningCalculator
 
 def print_result(result, expected):
